refactor(kotlin): Creation of monitored activities.
These classes, thanks to their life cycle, it is possible to monitor what the current activity is within the application and to be able to handle global events such as the absence of a network or a common event. The activities through interfaces have the possibility of communicating with the application (instance) and being able to carry out actions in a global and easy way.